+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-02-23
+### Added
+- Optional iterable object to the LinkedList constructor
+- LinkedList#add(value, position) method
+- LinkedList#remove(position) method
+### Removed
+- LinkedList.fromIterable()
+
 ## [0.1.0] - 2020-02-23
 ### Added
 - LinkedList collection